requirements.txt
tensorflow-gpu==1.8.0
keras
tqdm
numpy
coloredlogs
